+++ mne/io/brainvision/tests/test_brainvision.py
@@ -2,6 +2,7 @@
 from __future__ import print_function
 
 # Author: Teon Brooks <teon.brooks@gmail.com>
+#         Stefan Appelhoff <stefan.appelhoff@mailbox.org>
 #
 # License: BSD (3-clause)
 
@@ -20,6 +21,7 @@ from mne import pick_types, find_events
 from mne.io.constants import FIFF
 from mne.io import read_raw_fif, read_raw_brainvision
 from mne.io.tests.test_raw import _test_raw_reader
+from mne.io.write import DATE_NONE
 
 FILE = inspect.getfile(inspect.currentframe())
 data_dir = op.join(op.dirname(op.abspath(FILE)), 'data')
@@ -59,6 +61,24 @@ eog = ['HL', 'HR', 'Vb']
 event_id = {'Sync On': 5}
 
 warnings.simplefilter('always')
+
+
+def test_vmrk_meas_date():
+    """Test successful extraction of measurement date."""
+    # Test file that does have a specific date
+    raw = read_raw_brainvision(vhdr_path)
+    assert_allclose(raw.info['meas_date'], [1384359243, 794231])
+    assert '2013-11-13 16:14:03 GMT' in repr(raw.info)
+
+    # Test file with multiple dates ... we should only take the first
+    raw = read_raw_brainvision(vhdr_old_path)
+    assert_allclose(raw.info['meas_date'], [1184588560, 937453])
+    assert '2007-07-16 12:22:40 GMT' in repr(raw.info)
+
+    # Test files with no date, we should get DATE_NONE from mne.io.write
+    raw = read_raw_brainvision(vhdr_v2_path)
+    assert_allclose(raw.info['meas_date'], DATE_NONE)
+    assert 'unspecified' in repr(raw.info)
 
 
 def test_vhdr_codepage_ansi():